Refactor gRPC package to rpc and add tests for Assistant, Judge, Misc, Problem, Solution, and Testcase services
