+++ internal/interfaces/rpc/handler/solution.go
